iio: adc: ad7606: Rework to remove spi dependent code from core
The software mode configuration which strongly depends on spi was moved in
existent spi file.
This patch fixes core dependencies for spi.
