+++ drivers/iio/adc/ad7606.h
@@ -6,7 +6,36 @@
  */
 
 #ifndef IIO_ADC_AD7606_H_
+
+#include <linux/gpio/consumer.h>
+
 #define IIO_ADC_AD7606_H_
+
+#define AD760X_CHANNEL(num, mask_sep, mask_type, mask_all) {	\
+		.type = IIO_VOLTAGE,				\
+		.indexed = 1,					\
+		.channel = num,					\
+		.address = num,					\
+		.info_mask_separate = mask_sep,			\
+		.info_mask_shared_by_type = mask_type,		\
+		.info_mask_shared_by_all = mask_all,		\
+		.scan_index = num,				\
+		.scan_type = {					\
+			.sign = 's',				\
+			.realbits = 16,				\
+			.storagebits = 16,			\
+			.endianness = IIO_CPU,			\
+		},						\
+}
+
+#define AD7605_CHANNEL(num)				\
+	AD760X_CHANNEL(num, BIT(IIO_CHAN_INFO_RAW),	\
+		BIT(IIO_CHAN_INFO_SCALE), 0)
+
+#define AD7606_CHANNEL(num)				\
+	AD760X_CHANNEL(num, BIT(IIO_CHAN_INFO_RAW),	\
+		BIT(IIO_CHAN_INFO_SCALE),		\
+		BIT(IIO_CHAN_INFO_OVERSAMPLING_RATIO))
 
 /**
  * struct ad7606_chip_info - chip specific information
@@ -16,13 +45,6 @@
  *			oversampling ratios.
  * @oversampling_num	number of elements stored in oversampling_avail array
  * @os_req_reset	some devices require a reset to update oversampling
- * @spi_rd_wr_cmd	pointer to the function which calculates the spi address
- * @write_scale_sw	pointer to the function which writes the scale via spi
-			in software mode
- * @write_os_sw		pointer to the function which writes the os via spi
-			in software mode
- * @sw_mode_config:	pointer to a function which configured the device
- *			for software mode
  */
 struct ad7606_chip_info {
 	const struct iio_chan_spec	*channels;
@@ -30,10 +52,6 @@ struct ad7606_chip_info {
 	const unsigned int		*oversampling_avail;
 	unsigned int			oversampling_num;
 	bool				os_req_reset;
-	int (*spi_rd_wr_cmd)(int, char);
-	int (*write_scale_sw)(struct iio_dev *indio_dev, int, int);
-	int (*write_os_sw)(struct iio_dev *indio_dev, int);
-	int (*sw_mode_config)(struct iio_dev *indio_dev);
 };
 
 /**
@@ -103,10 +121,20 @@ struct ad7606_state {
 /**
  * struct ad7606_bus_ops - driver bus operations
  * @read_block		function pointer for reading blocks of data
+ * @sw_mode_config	pointer to a function which configured the device
+ *			for software mode
+ * @spi_rd_wr_cmd	pointer to the function which calculates the spi address
+ * @ad7606_reg_access	pointer to the function which handles the debug mode
  */
 struct ad7606_bus_ops {
 	/* more methods added in future? */
 	int (*read_block)(struct device *, int, void *);
+	int (*sw_mode_config)(struct iio_dev *indio_dev);
+	u16 (*spi_rd_wr_cmd)(int, char);
+	int (*ad7606_reg_access)(struct iio_dev *indio_dev,
+			     unsigned int reg,
+			     unsigned int writeval,
+			     unsigned int *readval);
 };
 
 int ad7606_probe(struct device *dev, int irq, void __iomem *base_address,
